object-detection-using-webcam/requirements.txt
ultralytics
opencv-python
roboflow
python-dotenv
lapx
